+++ src/PIL/BmpImagePlugin.py
@@ -484,7 +484,9 @@ def _save(
     if palette:
         fp.write(palette)
 
-    ImageFile._save(im, fp, [("raw", (0, 0) + im.size, 0, (rawmode, stride, -1))])
+    ImageFile._save(
+        im, fp, [ImageFile._Tile("raw", (0, 0) + im.size, 0, (rawmode, stride, -1))]
+    )
 
 
 #